+++ cherab/core/model/lineshape/zeeman.pyx
@@ -21,7 +21,7 @@
 from libc.math cimport sqrt
 from raysect.optical cimport Spectrum, Point3D, Vector3D
 
-from cherab.core.atomic cimport Line
+from cherab.core.atomic cimport Line, AtomicData
 from cherab.core.species cimport Species
 from cherab.core.plasma cimport Plasma
 from cherab.core.atomic.elements import hydrogen, deuterium, tritium, helium, helium3, beryllium, boron, carbon, nitrogen, oxygen, neon
@@ -51,6 +51,7 @@ cdef class ZeemanLineShapeModel(LineShapeModel):
     :param float wavelength: The rest wavelength for this emission line.
     :param Species target_species: The target plasma species that is emitting.
     :param Plasma plasma: The emitting plasma object.
+    :param AtomicData atomic_data: The atomic data provider.
     :param str polarisation: Leaves only :math:`\pi`-/:math:`\sigma`-polarised components:
                              "pi" - leave only :math:`\pi`-polarised components,
                              "sigma" - leave only :math:`\sigma`-polarised components,
@@ -59,9 +60,9 @@ cdef class ZeemanLineShapeModel(LineShapeModel):
                                     over the spectral bin. Default is None.
     """
 
-    def __init__(self, Line line, double wavelength, Species target_species, Plasma plasma, polarisation,
-                 Integrator1D integrator=None):
-        super().__init__(line, wavelength, target_species, plasma, integrator)
+    def __init__(self, Line line, double wavelength, Species target_species, Plasma plasma, AtomicData atomic_data,
+                 polarisation='no', Integrator1D integrator=None):
+        super().__init__(line, wavelength, target_species, plasma, atomic_data, integrator)
 
         self.polarisation = polarisation
 
@@ -94,15 +95,16 @@ cdef class ZeemanTriplet(ZeemanLineShapeModel):
     :param float wavelength: The rest wavelength for this emission line.
     :param Species target_species: The target plasma species that is emitting.
     :param Plasma plasma: The emitting plasma object.
+    :param AtomicData atomic_data: The atomic data provider.
     :param str polarisation: Leaves only :math:`\pi`-/:math:`\sigma`-polarised components:
                              "pi" - leave central component,
                              "sigma" - leave side components,
                              "no" - all components (default).
     """
 
-    def __init__(self, Line line, double wavelength, Species target_species, Plasma plasma, polarisation='no'):
+    def __init__(self, Line line, double wavelength, Species target_species, Plasma plasma, AtomicData atomic_data, polarisation='no'):
 
-        super().__init__(line, wavelength, target_species, plasma, polarisation)
+        super().__init__(line, wavelength, target_species, plasma, atomic_data, polarisation)
 
     @cython.boundscheck(False)
     @cython.wraparound(False)
@@ -174,9 +176,6 @@ cdef class ParametrisedZeemanTriplet(ZeemanLineShapeModel):
     :math:`\frac{W_{Zeeman}}{W_{Doppler}} = \beta T^{\gamma}`,
     where `T` is the species temperature in eV.
 
-    Call `show_supported_transitions()` to see the list of supported transitions and
-    default parameters of the model.
-
     For details see A. Blom and C. Jupén, Parametrisation of the Zeeman effect
     for hydrogen-like spectra in high-temperature plasmas,
     Plasma Phys. Control. Fusion 44 (2002) `1229-1241
@@ -186,70 +185,22 @@ cdef class ParametrisedZeemanTriplet(ZeemanLineShapeModel):
     :param float wavelength: The rest wavelength for this emission line.
     :param Species target_species: The target plasma species that is emitting.
     :param Plasma plasma: The emitting plasma object.
-    :param dict line_parameters: Alternative parameters of the model in the form
-                                 {line_i: (alpha_i, beta_i, gamma_i), ...}.
-                                 If None, the default model parameters will be used.
+    :param AtomicData atomic_data: The atomic data provider.
+    :param tuple line_parameters: Parameters of the model in the form (alpha, beta, gamma).
+                                  Default is None (will use `atomic_data.zeeman_triplet_parameters`).
     :param str polarisation: Leaves only :math:`\pi`-/:math:`\sigma`-polarised components:
                              "pi" - leave central component,
                              "sigma" - leave side components,
                              "no" - all components (default).
     """
 
-    LINE_PARAMETERS_DEFAULT = {  # alpha, beta, gamma parameters for selected lines
-        Line(hydrogen, 0, (3, 2)): (0.0402267, 0.3415, -0.5247),
-        Line(hydrogen, 0, (4, 2)): (0.0220724, 0.2837, -0.5346),
-        Line(deuterium, 0, (3, 2)): (0.0402068, 0.4384, -0.5015),
-        Line(deuterium, 0, (4, 2)): (0.0220610, 0.3702, -0.5132),
-        Line(helium3, 1, (4, 3)): (0.0205200, 1.4418, -0.4892),
-        Line(helium3, 1, (5, 3)): (0.0095879, 1.2576, -0.5001),
-        Line(helium3, 1, (6, 4)): (0.0401980, 0.8976, -0.4971),
-        Line(helium3, 1, (7, 4)): (0.0273538, 0.8529, -0.5039),
-        Line(helium, 1, (4, 3)): (0.0205206, 1.6118, -0.4838),
-        Line(helium, 1, (5, 3)): (0.0095879, 1.4294, -0.4975),
-        Line(helium, 1, (6, 4)): (0.0401955, 1.0058, -0.4918),
-        Line(helium, 1, (7, 4)): (0.0273521, 0.9563, -0.4981),
-        Line(beryllium, 3, (5, 4)): (0.0060354, 2.1245, -0.3190),
-        Line(beryllium, 3, (6, 5)): (0.0202754, 1.6538, -0.3192),
-        Line(beryllium, 3, (7, 5)): (0.0078966, 1.7017, -0.3348),
-        Line(beryllium, 3, (8, 6)): (0.0205025, 1.4581, -0.3450),
-        Line(boron, 4, (6, 5)): (0.0083423, 2.0519, -0.2960),
-        Line(boron, 4, (7, 6)): (0.0228379, 1.6546, -0.2941),
-        Line(boron, 4, (8, 6)): (0.0084065, 1.8041, -0.3177),
-        Line(boron, 4, (8, 7)): (0.0541883, 1.4128, -0.2966),
-        Line(boron, 4, (9, 7)): (0.0190781, 1.5440, -0.3211),
-        Line(boron, 4, (10, 8)): (0.0391914, 1.3569, -0.3252),
-        Line(carbon, 5, (6, 5)): (0.0040900, 2.4271, -0.2818),
-        Line(carbon, 5, (7, 6)): (0.0110398, 1.9785, -0.2816),
-        Line(carbon, 5, (8, 6)): (0.0040747, 2.1776, -0.3035),
-        Line(carbon, 5, (8, 7)): (0.0261405, 1.6689, -0.2815),
-        Line(carbon, 5, (9, 7)): (0.0092096, 1.8495, -0.3049),
-        Line(carbon, 5, (10, 8)): (0.0189020, 1.6191, -0.3078),
-        Line(carbon, 5, (11, 8)): (0.0110428, 1.6600, -0.3162),
-        Line(carbon, 5, (10, 9)): (0.0359009, 1.4464, -0.3104),
-        Line(nitrogen, 6, (7, 6)): (0.0060010, 2.4789, -0.2817),
-        Line(nitrogen, 6, (8, 7)): (0.0141271, 2.0249, -0.2762),
-        Line(nitrogen, 6, (9, 8)): (0.0300127, 1.7415, -0.2753),
-        Line(nitrogen, 6, (10, 8)): (0.0102089, 1.9464, -0.2975),
-        Line(nitrogen, 6, (11, 9)): (0.0193799, 1.7133, -0.2973),
-        Line(oxygen, 7, (8, 7)): (0.0083081, 2.4263, -0.2747),
-        Line(oxygen, 7, (9, 8)): (0.0176049, 2.0652, -0.2721),
-        Line(oxygen, 7, (10, 8)): (0.0059933, 2.3445, -0.2944),
-        Line(oxygen, 7, (10, 9)): (0.0343805, 1.8122, -0.2718),
-        Line(oxygen, 7, (11, 9)): (0.0113640, 2.0268, -0.2911),
-        Line(neon, 9, (9, 8)): (0.0072488, 2.8838, -0.2758),
-        Line(neon, 9, (10, 9)): (0.0141002, 2.4755, -0.2718),
-        Line(neon, 9, (11, 9)): (0.0046673, 2.8410, -0.2917),
-        Line(neon, 9, (11, 10)): (0.0257292, 2.1890, -0.2715)
-    }
+    def __init__(self, Line line, double wavelength, Species target_species, Plasma plasma, AtomicData atomic_data,
+                 tuple line_parameters=None, polarisation='no'):
 
-    def __init__(self, Line line, double wavelength, Species target_species, Plasma plasma, dict line_parameters=None, polarisation='no'):
-
-        super().__init__(line, wavelength, target_species, plasma, polarisation)
-
-        line_parameters = line_parameters or self.LINE_PARAMETERS_DEFAULT
+        super().__init__(line, wavelength, target_species, plasma, atomic_data, polarisation)
 
         try:
-            alpha, beta, gamma = line_parameters[self.line]
+            alpha, beta, gamma = line_parameters or self.atomic_data.zeeman_triplet_parameters(line)
             if alpha <= 0:
                 raise ValueError('Parameter alpha must be positive.')
             if beta < 0:
@@ -260,12 +211,6 @@ cdef class ParametrisedZeemanTriplet(ZeemanLineShapeModel):
 
         except KeyError:
             raise ValueError('Data for {} is not available.'.format(self.line))
-
-    @classmethod
-    def show_supported_transitions(cls):
-        """ Prints all supported transitions."""
-        for line, param in cls.LINE_PARAMETERS_DEFAULT.items():
-            print('{}: alpha={}, beta={}, gamma={}'.format(line, param[0], param[1], param[2]))
 
     @cython.boundscheck(False)
     @cython.wraparound(False)
@@ -337,9 +282,11 @@ cdef class ZeemanMultiplet(ZeemanLineShapeModel):
     :param float wavelength: The rest wavelength of the base emission line.
     :param Species target_species: The target plasma species that is emitting.
     :param Plasma plasma: The emitting plasma object.
+    :param AtomicData atomic_data: The atomic data provider.
     :param zeeman_structure: A ``ZeemanStructure`` object that provides wavelengths and ratios
                              of :math:`\pi`-/:math:`\sigma^{+}`-/:math:`\sigma^{-}`-polarised
                              components for any given magnetic field strength.
+                             Default is None (will use atomic_data.zeeman_structure).
     :param str polarisation: Leaves only :math:`\pi`-/:math:`\sigma`-polarised components:
                              "pi" - leave only :math:`\pi`-polarised components,
                              "sigma" - leave only :math:`\sigma`-polarised components,
@@ -347,12 +294,12 @@ cdef class ZeemanMultiplet(ZeemanLineShapeModel):
 
     """
 
-    def __init__(self, Line line, double wavelength, Species target_species, Plasma plasma,
-                 ZeemanStructure zeeman_structure, polarisation='no'):
+    def __init__(self, Line line, double wavelength, Species target_species, Plasma plasma, AtomicData atomic_data,
+                 ZeemanStructure zeeman_structure=None, polarisation='no'):
 
-        super().__init__(line, wavelength, target_species, plasma, polarisation)
+        super().__init__(line, wavelength, target_species, plasma, atomic_data, polarisation)
 
-        self._zeeman_structure = zeeman_structure
+        self._zeeman_structure = zeeman_structure or self.atomic_data.zeeman_structure(line)
 
     @cython.boundscheck(False)
     @cython.wraparound(False)